requirements.txt
os
pillow
datasets
clip
matplotlib
torch==2.7.0
torchvision==0.22.0
wget